Add initial source map support
By enabling the 'sourcemap' option, sourcemaps are generated.

The tests are very simple, though: They only check if source maps are
present, not if they are correct.

I'm new to grunt, so I couldn't find a way to test both with and without
source maps, so all tests have it enabled.

Moreover, there are no tests for in-sourcemaps.